fix(spectral12): independent-review remediation — P0 tie order, P1 gate recovery, P2 guards
- P0 (spectral_library_scale.cpp): the full-scoring path sorted by angle
  alone over bucket-push order, so equal/undefined angles ranked by bucket
  discovery instead of library order — diverging from matchSpectrum for
  cross-grid sentinel/zero entries. Sorts with the same (NaN->inf, entryIndex)
  key as the prescreened path; adds a cross-bucket regression fixture and
  tightens the equivalence assertions from 1e-12 margins to bit equality.
- P1: finishes the determinism-census/contract recovery — 20 uncovered ids
  receive family contract rows (the four Spectral 11.0 operators, three
  radiometric, three terrain, four InSAR, two registration, three io,
  seasonal breaks) and the census snapshot regenerates; both gates are green.
- P2 (rs_spectral_detection_operators.cpp): sample-count guard compares in
  size_t (int cast could wrap for >2^31 valid pixels).
- P2 (rs:spectral_spatial_fuse): radius bounded to [0, 128] in the schema and
  the driver (the O(pixels*(2r+1)^2) kernel has no interior cancellation
  point); honest FullRaster estimate 16 B/px.
- P3: streaming-test header no longer claims NaN scores for declared-NoData
  pixels (kernels NaN only non-finite inputs); prescreen=false now honors
  topK; kept-prefix bucket resamples are cached; bucket keys use the same
  sanitized grid semantics as membership; conditionProxy documents the
  orthogonal-start -> -1.0 edge; MatchIndex documents the Library lifetime
  requirement; ADR 0165 pin wording matches the delivered 55.

+++ src/operators/rs/rs_spectral_detection_operators.cpp
@@ -127,7 +127,7 @@ Json::Value runDetector( const std::string &kind, const Json::Value &params,
         if ( cemStats.count == 0 )
             throw RSOperatorError( ErrorCode::InvalidInputData, "No valid pixels found" );
         const int minSamples = SpectralCem::minSamplesRequired( bandCount, loading > 0.0 );
-        if ( static_cast<int>( cemStats.count ) < minSamples )
+        if ( cemStats.count < static_cast<size_t>( minSamples ) )
             throw RSOperatorError(
                 ErrorCode::InvalidInputData,
                 "CEM background is under-sampled: " + std::to_string( cemStats.count ) +
